+++ CHANGELOG.md
@@ -8,6 +8,10 @@ This version is compatible with [`ab4560900cf475ff515054bec0ca9a4491aca366`](htt
 - Improved `int_from_bytes` performance.
 - Improved `bigint_to_bytes` around 2 times faster.
 - Improved `int_to_bytes` performance.
+### Fixed
+- Fixed an issue where utility function `division`/`modulo` returned wrong values in some cases.
+### Added
+- Added `divmod` utility function.
 
 ## [1.0.5]
 This version is compatible with [`ab4560900cf475ff515054bec0ca9a4491aca366`](https://github.com/Chia-Network/clvm/tree/ab4560900cf475ff515054bec0ca9a4491aca366) of [clvm@0.9.7](https://github.com/Chia-Network/clvm)